break out context into it's own file, removing from ColumnsArea
